pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "copilot-skill-eval"
version = "0.1.0"
description = "A framework for evaluating how GitHub Copilot custom skills impact code generation quality"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "jinja2>=3.1",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/sayedihashimi/copilot-skill-eval"
Repository = "https://github.com/sayedihashimi/copilot-skill-eval"

[project.scripts]
skill-eval = "skill_eval.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
skill_eval = ["templates/*"]
